+++ tests/covalent_braket_plugin_tests/braket_test.py
@@ -28,8 +28,10 @@ from unittest.mock import AsyncMock, MagicMock
 
 import cloudpickle
 import pytest
+from botocore.exceptions import ClientError
+from covalent._shared_files.exceptions import TaskCancelledError
 
-from covalent_braket_plugin.braket import BraketExecutor
+from covalent_braket_plugin.braket import BRAKET_JOB_NAME, BraketExecutor
 
 MOCK_CREDENTIALS = "mock_credentials"
 MOCK_PROFILE = "mock_profile"
@@ -102,6 +104,8 @@ async def test_run(braket_executor, mocker):
         return_value=query_result_async_mock,
     )
     query_result_mock.return_value = "result", "", ""
+    braket_executor.get_cancel_requested = AsyncMock(return_value=False)
+    braket_executor.set_job_handle = AsyncMock()
 
     await braket_executor.run(
         function=mock_func, args=[], kwargs={"x": 1}, task_metadata=task_metadata
@@ -112,6 +116,68 @@ async def test_run(braket_executor, mocker):
         mock_func, [], {"x": 1}, {"image_tag": "mock_dispatch_id-1"}
     )
     submit_task_mock.assert_called_once()
+
+
+@pytest.mark.asyncio
+async def test_run_cancel_before_upload(braket_executor, mocker):
+    """Test the run method when cancel is requested before uploading to S3"""
+
+    def mock_func(x):
+        return x
+
+    dispatch_id = "mock_dispatch_id"
+    node_id = 1
+    results_dir = "/tmp"
+    task_metadata = {"dispatch_id": dispatch_id, "node_id": node_id, "results_dir": results_dir}
+    MOCK_BRAKET_JOB_NAME = BRAKET_JOB_NAME.format(dispatch_id=dispatch_id, node_id=node_id)
+    validate_creds_mock = mocker.patch(
+        "covalent_braket_plugin.braket.BraketExecutor._validate_credentials"
+    )
+    braket_executor.get_cancel_requested = AsyncMock(return_value=True)
+
+    with pytest.raises(TaskCancelledError) as exception:
+        await braket_executor.run(
+            function=mock_func, args=[], kwargs={"x": 1}, task_metadata=task_metadata
+        )
+        assert exception == f"AWS Batch job {MOCK_BRAKET_JOB_NAME} requested to be cancelled"
+
+    validate_creds_mock.assert_called_once()
+    braket_executor.get_cancel_requested.assert_called_once()
+
+
+@pytest.mark.asyncio
+async def test_run_cancel_before_submit_task(braket_executor, mocker):
+    """Test the run method when cancel is requested before submitting task to AWS Braket."""
+
+    def mock_func(x):
+        return x
+
+    dispatch_id = "mock_dispatch_id"
+    node_id = 1
+    results_dir = "/tmp"
+    task_metadata = {"dispatch_id": dispatch_id, "node_id": node_id, "results_dir": results_dir}
+    MOCK_BRAKET_JOB_NAME = BRAKET_JOB_NAME.format(dispatch_id=dispatch_id, node_id=node_id)
+
+    mocker.patch("covalent_braket_plugin.braket.boto3")
+    validate_creds_mock = mocker.patch(
+        "covalent_braket_plugin.braket.BraketExecutor._validate_credentials"
+    )
+    upload_task_mock = mocker.patch(
+        "covalent_braket_plugin.braket.BraketExecutor._upload_task", return_value=AsyncMock()
+    )
+    braket_executor.get_cancel_requested = AsyncMock(side_effect=[False, True])
+
+    with pytest.raises(TaskCancelledError) as exception:
+        await braket_executor.run(
+            function=mock_func, args=[], kwargs={"x": 1}, task_metadata=task_metadata
+        )
+        assert exception == f"AWS Batch job {MOCK_BRAKET_JOB_NAME} requested to be cancelled"
+
+    validate_creds_mock.assert_called_once()
+    upload_task_mock.assert_called_once_with(
+        mock_func, [], {"x": 1}, {"image_tag": "mock_dispatch_id-1"}
+    )
+    assert braket_executor.get_cancel_requested.call_count == 2
 
 
 @pytest.mark.asyncio
@@ -218,5 +284,50 @@ async def test_query_result(braket_executor, mocker):
 
 
 @pytest.mark.asyncio
-async def test_stubs(braket_executor):
-    await braket_executor.cancel()
+async def test_cancel_braket_task(braket_executor, mocker):
+    boto3_mock = mocker.patch("covalent_braket_plugin.braket.boto3")
+    boto3_client_mock = boto3_mock.Session().client()
+    boto3_client_mock.cancel_quantum_task.return_value = {"status": "CANCELLED"}
+    mock_arn = (
+        "arn:aws:braket:us-west-2:123456789012:quantum-task/01234567-89ab-cdef-0123-456789abcdef"
+    )
+    mock_dispatch_id = "abcdef"
+    mock_node_id = 0
+    mock_task_metadata = {"dispatch_id": mock_dispatch_id, "node_id": mock_node_id}
+
+    is_cancelled = await braket_executor.cancel(
+        task_metadata=mock_task_metadata, job_handle=mock_arn
+    )
+
+    assert is_cancelled is True
+    assert boto3_client_mock.cancel_quantum_task.called_once_with(quantumTaskArn=mock_arn)
+
+
+@pytest.mark.asyncio
+async def test_cancel_failed_braket_task(braket_executor, mocker):
+    boto3_mock = mocker.patch("covalent_braket_plugin.braket.boto3")
+    boto3_client_mock = boto3_mock.Session().client()
+    boto3_client_mock.cancel_quantum_task.return_value = {"status": "CANCELLED"}
+    mock_arn = (
+        "arn:aws:braket:us-west-2:123456789012:quantum-task/01234567-89ab-cdef-0123-456789abcdef"
+    )
+    mock_dispatch_id = "abcdef"
+    mock_node_id = 0
+    mock_task_metadata = {"dispatch_id": mock_dispatch_id, "node_id": mock_node_id}
+    mock_error = ClientError(
+        operation_name="TerminateJob",
+        error_response={
+            "Error": {
+                "Code": "UnreachableHostException",
+                "Message": 'Could not connect to the endpoint URL: "https://braket.us-east-1.amazonaws.com/v1/quantum-task"',
+            }
+        },
+    )
+    boto3_client_mock.cancel_quantum_task.side_effect = mock_error
+
+    is_cancelled = await braket_executor.cancel(
+        task_metadata=mock_task_metadata, job_handle=mock_arn
+    )
+
+    assert is_cancelled is False
+    assert boto3_client_mock.cancel_quantum_task.called_once_with(mock_arn)